requirements.txt
# Django
Django>=5.2.1

# Faker for generating fake data
Faker>=19.3.0

# Requests for fetching images from Unsplash
requests>=2.31.0

# Python-dotenv for loading environment variables
python-dotenv>=1.0.0

# Mutagen for ID3 metadata extraction
mutagen>=1.47.0

djangorestframework~=3.16.0
dotenv~=0.9.9
reportlab~=4.4.1